+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 Note: Can be used with `oxsecurity/megalinter@beta` in your GitHub Action mega-linter.yml file, or with `oxsecurity/megalinter:beta` docker image
 
 - Core
+  - New LLM Advisor: call external LLMs to get hints to solve linter errors, available in:
+    - Console Reporter
+    - Text Reporter
+    - Git platforms PR/MR comments Reporter
   - Add copilot instructions for GitHub Copilot
 
 - New linters
@@ -19,6 +23,7 @@ Note: Can be used with `oxsecurity/megalinter@beta` in your GitHub Action mega-l
 
 - Linters enhancements
   - PHP-CS-Fixer is able to run on PHP 8.4 without error (change default configuration) by @llaville
+  - [cspell](https://megalinter.io/latest/descriptors/spell_cspell/): Filter output lines that do not contain found issues
 
 - Fixes
 
@@ -27,6 +32,7 @@ Note: Can be used with `oxsecurity/megalinter@beta` in your GitHub Action mega-l
   - Markdown summary reporter:
     - Write a file for Github integration if GITHUB_STEP_SUMMARY is set
     - Truncate less linter output lines
+  - Text reporter: Change the output file names to put the linter name first, then the status
 
 - Doc
   - Update documentation in all megalinter descriptor files to improve accuracy and consistency